main: move status methods to status*.go
